gymnasium/requirements.txt
numpy==2.3.5
gymnasium[classic-control]==1.2.2
pyzmq==27.1.0
matplotlib==3.10.8